Add comprehensive test suite for logging functionality
- Introduce multiple test cases to validate logger behavior, including theme selection, color handling, and exception formatting.
- Implement helper functions for streamlined logger setup in tests.
- Ensure coverage for various logging scenarios, including custom themes and environment variable effects on color output.
- Remove outdated test file to maintain clarity and focus in the test suite.
